requirements.txt
aniso8601==9.0.1
bcrypt==4.0.1
beautifulsoup4==4.12.0
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql==0.0.3
mysql-connector-python==8.0.32
mysqlclient==2.1.1
protobuf==3.20.3
psutil==5.9.4
# pypiwin32==223
python-dotenv==1.0.0
pytz==2022.7.1
# pywin32==305
requests==2.28.1
six==1.16.0
soupsieve==2.4
SQLAlchemy==2.0.8
typing_extensions==4.5.0
urllib3==1.26.12
Werkzeug==2.2.2
winshell==0.6
